Added int to avoid deprecation warning message (DeprecationWarning: using a non-integer number instead of an integer will result in an error in the future)
git-svn-id: http://svn.astro.pic.es/PAUdm/branches/bt_davide@1253 62b34b79-d161-4497-8a8a-8b88d6ba3201

+++ paudm/tools/common/objects.py
@@ -992,8 +992,8 @@ class Catalogue(object):
             y = -2048. * n * math.sqrt(1-n*math.sin(b))*math.sin(l) + 2047.5
             
             # look up the value of that pixel in the dust map
-            x_int = math.floor(x+0.5)
-            y_int = math.floor(y+0.5)
+            x_int = int(math.floor(x+0.5))
+            y_int = int(math.floor(y+0.5))
             if n == 1:
                 extinction = SFD_north[y_int-1, x_int-1];
             else:
@@ -1081,8 +1081,8 @@ class Catalogue(object):
             y = -2048. * n * math.sqrt(1-n*math.sin(b))*math.sin(l) + 2047.5
             
             # look up the value of that pixel in the dust map
-            x_int = math.floor(x+0.5)
-            y_int = math.floor(y+0.5)
+            x_int = int(math.floor(x+0.5))
+            y_int = int(math.floor(y+0.5))
             if n == 1:
                 extinction = SFD_north[y_int-1, x_int-1];
             else:
